feat(zink): import keccak in non-wasm targets
