+++ Cargo.toml
@@ -100,6 +100,9 @@ path = "zink/src/lib.rs"
 paste.workspace = true
 zink-codegen.workspace = true
 
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+tiny-keccak.workspace = true
+
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
 anyhow.workspace = true
 filetests.workspace = true
