+++ zink/codegen/Cargo.toml
@@ -18,5 +18,4 @@ hex.workspace = true
 proc-macro2.workspace = true
 quote.workspace = true
 syn.workspace = true
-tiny-keccak.workspace = true
 zabi = { workspace = true, features = [ "hex", "syn" ] }
